Consolidate shader binding definitions and some more code cleanup

+++ bender/bender-base/app/src/main/jni/vulkan_main.cc
@@ -223,24 +223,22 @@ void updateUniformBuffer(uint32_t frameIndex) {
 void createDescriptorPool() {
   std::array<VkDescriptorPoolSize, 4> poolSizes = {};
   poolSizes[0].type = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-  poolSizes[0].descriptorCount = static_cast<uint32_t>(device->getDisplayImagesSize());
+  poolSizes[0].descriptorCount = device->getDisplayImagesSize();
   poolSizes[1].type = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-  poolSizes[1].descriptorCount = static_cast<uint32_t>(device->getDisplayImagesSize());
+  poolSizes[1].descriptorCount = device->getDisplayImagesSize();
   poolSizes[2].type = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-  poolSizes[2].descriptorCount = static_cast<uint32_t>(device->getDisplayImagesSize());
+  poolSizes[2].descriptorCount = device->getDisplayImagesSize();
   poolSizes[3].type = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-  poolSizes[3].descriptorCount = static_cast<uint32_t>(device->getDisplayImagesSize());
-
+  poolSizes[3].descriptorCount = device->getDisplayImagesSize();
 
   VkDescriptorPoolCreateInfo poolInfo = {};
   poolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-  poolInfo.poolSizeCount = static_cast<uint32_t>(poolSizes.size());
+  poolInfo.poolSizeCount = poolSizes.size();
   poolInfo.pPoolSizes = poolSizes.data();
-  poolInfo.maxSets = static_cast<uint32_t>(device->getDisplayImagesSize());
+  poolInfo.maxSets = device->getDisplayImagesSize();
 
   CALL_VK(vkCreateDescriptorPool(device->getDevice(), &poolInfo, nullptr, &descriptorPool_));
 }
-
 
 void createDescriptorSets() {
   std::vector<VkDescriptorSetLayout> layouts(device->getDisplayImagesSize(), mesh->getDescriptorSetLayout());
